# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UAS Laravel API Obat Documentation",
        "description": "Dibuat oleh Luthfi Nur Ramadhan ✨ - 2142430",
        "contact": {
            "email": "luthfiramadhan.lr55@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "description": "login",
                "operationId": "auth_login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Login Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "description": "Register",
                "operationId": "auth_register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Register",
                                    "data": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/obat": {
            "get": {
                "tags": [
                    "Obat"
                ],
                "summary": "",
                "description": "Get all Data",
                "operationId": "obat_index",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per Page value is number. ex : ?per_page=10",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page value is number. ex : ?page=10",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort value is string with rule column-name:order. ex : ?sort=id:asc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "where",
                        "in": "query",
                        "description": "Where value is object. ex : ?where{'name':'Izuchii','Izu':'2002-11-13'}",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Count value is boolean. ex : ?count=true",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": true,
                                    "message": "Get Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Obat"
                ],
                "summary": "",
                "description": "Insert Data",
                "operationId": "obat_store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "tipe": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "text"
                                    },
                                    "produksi": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "kadaluarsa": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Insert Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/obat/{id}": {
            "get": {
                "tags": [
                    "Obat"
                ],
                "summary": "",
                "description": "Get data by id",
                "operationId": "Obat_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": true,
                                    "message": "Get Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Obat"
                ],
                "summary": "",
                "description": "Update Data",
                "operationId": "obat_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "tipe": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "text"
                                    },
                                    "produksi": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "kadaluarsa": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Update Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Obat"
                ],
                "summary": "",
                "description": "Delete data",
                "operationId": "obat_destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": true,
                                    "message": "Delete Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}